Add Possibility to Stop CRUD Executions on Before Hooks
If the callback of the before hooks contain a string or an object, the
execution of the running CRUD operation will be stoped and an
appropriated error thrown.

+ model.prototype.errorFor
Added the #errorFor method to each model. This method returns the error
for the provided key and additonally registers the appropriated change
events.